+++ frontend/app/companies/[id]/(waiting-admin)/waiting-dashboard/settings/page.tsx
@@ -1,4 +1,4 @@
-import { WaitingDashboardSettings } from "@/components/companies/WaitingDashboardSettings";
+import { WaitingDashboardSettings } from "@/components/companies/waitings/WaitingDashboardSettings";
 
 export default function WaitingDashboardSettingsPage() {
     return (
