+++ frontend/components/companies/waitings/WaitingCard.tsx
