+++ frontend/components/companies/waitings/WaitingListCard.tsx
